requirements.txt
streamlit
pandas
feedparser
fpdf
python-nmap
groq